Refactor and add description
