+++ client/apps/settings/client/curations/gucci/test/series.test.js
@@ -96,50 +96,4 @@ describe('Series Admin', () => {
     expect(props.onChange.mock.calls[0][0]).toMatch('partner_logo_secondary')
     expect(props.onChange.mock.calls[0][1]).toMatch('http://new-logo.jpg')
   })
-
-  it('Updates metadata inputs', () => {
-    const component = mount(
-      <SeriesAdmin {...props} />
-    )
-    const inputs = component.find('input')
-    const textarea = component.find('textarea')
-    inputs.at(3).simulate('change', { target: { value: 'Social Title' } })
-    inputs.at(4).simulate('change', { target: { value: 'Email Title' } })
-    inputs.at(5).simulate('change', { target: { value: 'Email Author' } })
-    inputs.at(6).simulate('change', { target: { value: 'Email Tags' } })
-    inputs.at(7).simulate('change', { target: { value: 'Keywords' } })
-    textarea.at(0).simulate('change', { target: { value: 'Social Description' } })
-
-    expect(props.onChange.mock.calls[0][0]).toMatch('social_title')
-    expect(props.onChange.mock.calls[0][1]).toMatch('Social Title')
-    expect(props.onChange.mock.calls[1][0]).toMatch('email_title')
-    expect(props.onChange.mock.calls[1][1]).toMatch('Email Title')
-    expect(props.onChange.mock.calls[2][0]).toMatch('email_author')
-    expect(props.onChange.mock.calls[2][1]).toMatch('Email Author')
-    expect(props.onChange.mock.calls[3][0]).toMatch('email_tags')
-    expect(props.onChange.mock.calls[3][1]).toMatch('Email Tags')
-    expect(props.onChange.mock.calls[4][0]).toMatch('keywords')
-    expect(props.onChange.mock.calls[4][1]).toMatch('Keywords')
-    expect(props.onChange.mock.calls[5][0]).toMatch('social_description')
-    expect(props.onChange.mock.calls[5][1]).toMatch('Social Description')
-  })
-
-  it('Updates metadata images', () => {
-    const component = mount(
-      <SeriesAdmin {...props} />
-    )
-    const thumbnailInput = component.find(ImageUpload).at(2).node
-    thumbnailInput.props.onChange(thumbnailInput.props.name, 'thumbnailImage.jpg')
-    const socialInput = component.find(ImageUpload).at(3).node
-    socialInput.props.onChange(socialInput.props.name, 'socialImage.jpg')
-    const emailInput = component.find(ImageUpload).at(4).node
-    emailInput.props.onChange(emailInput.props.name, 'emailImage.jpg')
-
-    expect(props.onChange.mock.calls[0][0]).toMatch('thumbnail_image')
-    expect(props.onChange.mock.calls[0][1]).toMatch('thumbnailImage.jpg')
-    expect(props.onChange.mock.calls[1][0]).toMatch('social_image')
-    expect(props.onChange.mock.calls[1][1]).toMatch('socialImage.jpg')
-    expect(props.onChange.mock.calls[2][0]).toMatch('email_image')
-    expect(props.onChange.mock.calls[2][1]).toMatch('emailImage.jpg')
-  })
 })
